🔴🚚 api: Move comment-by-jury route to session-management context
