+++ api/src/certification/session-management/infrastructure/repositories/course-assessment-result-repository.js
