+++ api/src/certification/session-management/infrastructure/serializers/jury-comment-serializer.js
@@ -1,6 +1,6 @@
 import jsonapiSerializer from 'jsonapi-serializer';
 
-import { _ } from '../../../../../shared/infrastructure/utils/lodash-utils.js';
+import { _ } from '../../../../shared/infrastructure/utils/lodash-utils.js';
 const { Deserializer } = jsonapiSerializer;
 
 const deserialize = async function (payload) {
