+++ api/tests/certification/session-management/integration/infrastructure/repositories/course-assessment-result-repository_test.js
@@ -1,4 +1,4 @@
-import * as courseAssessmentResultRepository from '../../../../../../src/certification/course/infrastructure/repositories/course-assessment-result-repository.js';
+import * as courseAssessmentResultRepository from '../../../../../../src/certification/session-management/infrastructure/repositories/course-assessment-result-repository.js';
 import { AutoJuryCommentKeys } from '../../../../../../src/certification/shared/domain/models/JuryComment.js';
 import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
 import { AssessmentResult } from '../../../../../../src/shared/domain/models/AssessmentResult.js';
